+++ src/Message/FetchSubscriptionsRequest.php
@@ -71,8 +71,6 @@ use stdClass;
  */
 class FetchSubscriptionsRequest extends FetchTransactionsRequest
 {
-    const DEFAULT_PAGE_SIZE = 10000;
-
     /**
      * @return string
      */
@@ -81,56 +79,10 @@ class FetchSubscriptionsRequest extends FetchTransactionsRequest
         return self::$SUBSCRIPTION_OBJECT;
     }
 
-    /**
-     * Get the number of records to return per call
-     *
-     * @return null|int
-     */
-    public function getPageSize()
-    {
-        return $this->getParameter('pageSize');
-    }
-
-    /**
-     * Set the number of records to return per call
-     *
-     * @param int $value
-     * @return static
-     */
-    public function setPageSize($value)
-    {
-        return $this->setParameter('pageSize', $value);
-    }
-
-    /**
-     * Get the page to return. Starts at 0.
-     * For example, if pageSize is 10 and page is 0, returns the first 10
-     * results. If page is 1, returns the second 10 results.
-     *
-     * @return null|int
-     */
-    public function getPage()
-    {
-        return $this->getParameter('page');
-    }
-
-    /**
-     * Set the page to return. Starts at 0.
-     * For example, if pageSize is 10 and page is 0, returns the first 10
-     * results. If page is 1, returns the second 10 results.
-     *
-     * @param int $value
-     * @return static
-     */
-    public function setPage($value)
-    {
-        return $this->setParameter('page', $value);
-    }
-
     public function getData()
     {
         $data = parent::getData();
-        // foir some reason, page and pageSize are not optional for autobills
+        // for some reason, page and pageSize are not optional for autobills
         $data['page'] = $this->getPage() ?: 0;
         $data['pageSize'] = $this->getPageSize() ?: self::DEFAULT_PAGE_SIZE;
         return $data;
