+++ src/Message/FetchTransactionsRequest.php
@@ -72,10 +72,8 @@ use Omnipay\Common\Exception\InvalidRequestException;
  *
  * </code>
  */
-class FetchTransactionsRequest extends AbstractRequest
+class FetchTransactionsRequest extends AbstractPageableRequest
 {
-    const DEFAULT_PAGE_SIZE = 1000;
-
     /**
      * The name of the function to be called in Vindicia's API
      *
@@ -92,52 +90,6 @@ class FetchTransactionsRequest extends AbstractRequest
     protected function getObject()
     {
         return self::$TRANSACTION_OBJECT;
-    }
-
-    /**
-     * Get the number of records to return per call
-     *
-     * @return null|int
-     */
-    public function getPageSize()
-    {
-        return $this->getParameter('pageSize');
-    }
-
-    /**
-     * Set the number of records to return per call
-     *
-     * @param int $value
-     * @return static
-     */
-    public function setPageSize($value)
-    {
-        return $this->setParameter('pageSize', $value);
-    }
-
-    /**
-     * Get the page to return. Starts at 0.
-     * For example, if pageSize is 10 and page is 0, returns the first 10
-     * results. If page is 1, returns the second 10 results.
-     *
-     * @return null|int
-     */
-    public function getPage()
-    {
-        return $this->getParameter('page');
-    }
-
-    /**
-     * Set the page to return. Starts at 0.
-     * For example, if pageSize is 10 and page is 0, returns the first 10
-     * results. If page is 1, returns the second 10 results.
-     *
-     * @param int $value
-     * @return static
-     */
-    public function setPage($value)
-    {
-        return $this->setParameter('page', $value);
     }
 
     public function getData()
